+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 📝 docs(lib)-add comprehensive documentation for hcaptcha(pr [#25])
 - ♻️ refactor(lib)-enable error module(pr [#27])
 - Add-Captval-trait-for-validation(pr [#31])
+- ♻️ refactor(hcaptcha)-clean up unused code and comments(pr [#32])
 
 ### Security
 
@@ -81,5 +82,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#30]: https://github.com/jerus-org/captval/pull/30
 [#30]: https://github.com/jerus-org/captval/pull/30
 [#31]: https://github.com/jerus-org/captval/pull/31
+[#32]: https://github.com/jerus-org/captval/pull/32
 [Unreleased]: https://github.com/jerus-org/captval/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/jerus-org/captval/releases/tag/v0.1.0